+++ changelog.md
@@ -175,6 +175,7 @@ Codion Change Log
 - EntityTablePanel bug fixed, now overrides updateUI() in order to update hidden components.
 - EntityDialogs, add button caption now Add instead of Insert.
 - EntityFieldFactory renamed EntityConditionFieldFactory, now public.
+- EntityComboBox now displays a wait cursor while the model is refreshing.
 ### is.codion.swing.framework.ui.tools
 - DomainGeneratorPanel.DEFAULT_USERNAME configuration value added.
 - DomainGeneratorPanel.DEFAULT_USERNAME renamed DEFAULT_USER, can include password.